+++ tests/unit/conftest.py
@@ -1,5 +1,1 @@
-# Import fixtures from the fixtures modules to make them available to pytest
-
-from .fixtures.guilds import guild_roles, test_guild
-from .fixtures.binds import entire_group_bind
-from .fixtures.groups import test_group, group_rolesets, member_roleset
+from .fixtures import *  # pylint: disable=wildcard-import, unused-wildcard-import # makes fixtures available to all tests
