+++ tests/unit/utils/__init__.py
@@ -1,3 +1,3 @@
 from .snowflakes import generate_snowflake
-from .nickname_helpers import NicknameTestData, nickname_formatter
+from .nickname_helpers import nickname_formatter
 from .enum_helpers import filter_enum_list
